+++ framework/model/src/main/java/is/codion/framework/model/DefaultEntityApplicationModel.java
@@ -34,12 +34,12 @@ import static java.util.Objects.requireNonNull;
 
 /**
  * A central application model class.
- * @param <M> the type of {@link DefaultEntityModel} this application model is based on
- * @param <E> the type of {@link AbstractEntityEditModel} used by this {@link EntityModel}
+ * @param <M> the type of {@link EntityModel} this application model is based on
+ * @param <E> the type of {@link EntityEditModel} used by this {@link EntityModel}
  * @param <T> the type of {@link EntityTableModel} used by this {@link EntityModel}
  */
-public class DefaultEntityApplicationModel<M extends DefaultEntityModel<M, E, T>,
-				E extends AbstractEntityEditModel, T extends EntityTableModel<E>> implements EntityApplicationModel<M, E, T> {
+public class DefaultEntityApplicationModel<M extends EntityModel<M, E, T>,
+				E extends EntityEditModel, T extends EntityTableModel<E>> implements EntityApplicationModel<M, E, T> {
 
 	private final EntityConnectionProvider connectionProvider;
 	private final Version version;
@@ -104,8 +104,8 @@ public class DefaultEntityApplicationModel<M extends DefaultEntityModel<M, E, T>
 		}
 	}
 
-	private final class DefaultEntityModels<M extends DefaultEntityModel<M, E, T>,
-					E extends AbstractEntityEditModel, T extends EntityTableModel<E>> implements EntityModels<M, E, T> {
+	private final class DefaultEntityModels<M extends EntityModel<M, E, T>,
+					E extends EntityEditModel, T extends EntityTableModel<E>> implements EntityModels<M, E, T> {
 
 		private final List<M> entityModels = new ArrayList<>();