+++ changelog.md
@@ -32,6 +32,7 @@ Codion Change Log
 ### is.codion.framework.model
 - DefaultForeignKeyConditionModel now links search models to their respective operands.
 - ForeignKeyDetailModelLink, configuration methods renamed along with associated properties.
+- DefaultEntityModel, type parameters now refer to EntityModel and EntityEditModel instead of DefaultEntityModel and AbstractEntityEditModel.
 ### is.codion.swing.framework.model
 - DefaultSwingForeignKeyConditionModel now links search and combo box models to their respective operands.
 ### is.codion.swing.framework.ui